Prototype/Code/backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
SQLAlchemy==2.0.34
pymysql==1.1.1
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.9.0
pydantic-settings==2.5.2
email-validator==2.2.0
